# [REVIEW] Speed-up test_kmeans.py