--- a/python/cuml/test/test_kmeans.py
+++ b/python/cuml/test/test_kmeans.py
@@ -31,13 +31,9 @@ dataset_names = ['blobs', 'noisy_circles', 'noisy_moons', 'varied', 'aniso']
 
 SCORE_EPS = 0.06
 
-
-@pytest.mark.xfail
-@pytest.mark.parametrize('random_state', [i for i in range(10)])
-def test_n_init_cluster_consistency(random_state):
-
+@pytest.fixture
+def get_data():
     cluster_std = 1.0
-
     nrows = 1000
     ncols = 50
     nclusters = 8
@@ -48,6 +44,13 @@ def test_n_init_cluster_consistency(random_state):
                       cluster_std=cluster_std,
                       shuffle=False,
                       random_state=0)
+    return X, y
+
+@pytest.mark.xfail
+@pytest.mark.parametrize('random_state', [i for i in range(10)])
+def test_n_init_cluster_consistency(random_state):
+
+    X, y = get_data()
 
     cuml_kmeans = cuml.KMeans(verbose=0, init="k-means++",
                               n_clusters=nclusters,
@@ -201,8 +204,6 @@ def test_kmeans_sklearn_comparison_default(name, nrows):
     assert sk_score - 1e-2 <= cu_score <= sk_score + 1e-2
 
 
-@pytest.mark.parametrize('n_rows', [unit_param(100),
-                                    stress_param(500000)])
 @pytest.mark.parametrize('n_clusters', [unit_param(10),
                                         unit_param(100),
                                         stress_param(1000)])
@@ -212,7 +213,7 @@ def test_kmeans_sklearn_comparison_default(name, nrows):
 @pytest.mark.parametrize('init', ['k-means||',
                                   'random',
                                   'preset'])
-def test_all_kmeans_params(n_rows, n_clusters, max_iter, init,
+def test_all_kmeans_params(n_clusters, max_iter, init,
                            oversampling_factor, max_samples_per_batch):
 
     np.random.seed(0)