https://github.com/rapidsai/cuml/issues/2007 

Original Timing
`1174 passed, 355 skipped, 10 xfailed, 54 warnings in 52.22s` 

Showing the 20 most timing consuming tests:
<details>

```
1.85s call     cuml/test/test_kmeans.py::test_n_init_cluster_consistency[0]
0.26s call     cuml/test/test_kmeans.py::test_n_init_cluster_consistency[7]
0.25s call     cuml/test/test_kmeans.py::test_kmeans_sequential_plus_plus_init[42-5-50-10000]
0.25s call     cuml/test/test_kmeans.py::test_n_init_cluster_consistency[9]
0.25s call     cuml/test/test_kmeans.py::test_n_init_cluster_consistency[2]
0.25s call     cuml/test/test_kmeans.py::test_n_init_cluster_consistency[3]
0.25s call     cuml/test/test_kmeans.py::test_kmeans_sequential_plus_plus_init[16-5-50-10000]
0.24s call     cuml/test/test_kmeans.py::test_n_init_cluster_consistency[1]
0.24s call     cuml/test/test_kmeans.py::test_n_init_cluster_consistency[8]
0.24s call     cuml/test/test_kmeans.py::test_n_init_cluster_consistency[6]
0.24s call     cuml/test/test_kmeans.py::test_n_init_cluster_consistency[4]
0.24s call     cuml/test/test_kmeans.py::test_kmeans_sequential_plus_plus_init[48-5-50-10000]
0.24s call     cuml/test/test_kmeans.py::test_kmeans_sequential_plus_plus_init[49-5-50-10000]
0.24s call     cuml/test/test_kmeans.py::test_n_init_cluster_consistency[5]
0.24s call     cuml/test/test_kmeans.py::test_kmeans_sequential_plus_plus_init[35-5-50-10000]
0.24s call     cuml/test/test_kmeans.py::test_kmeans_sequential_plus_plus_init[23-5-50-10000]
0.24s call     cuml/test/test_kmeans.py::test_kmeans_sequential_plus_plus_init[33-5-50-10000]
0.24s call     cuml/test/test_kmeans.py::test_kmeans_sequential_plus_plus_init[34-5-50-10000]
0.24s call     cuml/test/test_kmeans.py::test_kmeans_sequential_plus_plus_init[29-5-50-10000]
0.24s call     cuml/test/test_kmeans.py::test_kmeans_sequential_plus_plus_init[31-5-50-10000]
```

</details>

After changes
`174 passed, 129 skipped, 54 warnings in 5.05s`

Top 20 most time consuming tests:
<details>

```
1.36s call     cuml/test/test_kmeans.py::test_kmeans_sklearn_comparison[1000-blobs]
0.30s call     cuml/test/test_kmeans.py::test_score[5-10-500]
0.09s call     cuml/test/test_kmeans.py::test_score[5-30-500]
0.06s call     cuml/test/test_kmeans.py::test_kmeans_sklearn_comparison[1000-aniso]
0.06s call     cuml/test/test_kmeans.py::test_kmeans_sklearn_comparison[1000-varied]
0.05s call     cuml/test/test_kmeans.py::test_kmeans_sklearn_comparison[1000-noisy_circles]
0.05s call     cuml/test/test_kmeans.py::test_kmeans_sklearn_comparison[1000-noisy_moons]
0.05s call     cuml/test/test_kmeans.py::test_kmeans_sklearn_comparison_default[500-noisy_circles]
0.04s call     cuml/test/test_kmeans.py::test_kmeans_sklearn_comparison_default[500-aniso]
0.04s call     cuml/test/test_kmeans.py::test_all_kmeans_params[k-means||-32-1.0-100-100]
0.04s call     cuml/test/test_kmeans.py::test_all_kmeans_params[k-means||-32-1.0-500-100]
0.04s call     cuml/test/test_kmeans.py::test_all_kmeans_params[k-means||-32-1.0-1000-100]
0.04s call     cuml/test/test_kmeans.py::test_all_kmeans_params[k-means||-32-0.5-1000-100]
0.04s call     cuml/test/test_kmeans.py::test_all_kmeans_params[k-means||-32-0.5-500-100]
0.04s call     cuml/test/test_kmeans.py::test_all_kmeans_params[k-means||-32-0.5-100-100]
0.04s call     cuml/test/test_kmeans.py::test_all_kmeans_params[k-means||-32-0.5-1000-10]
0.04s call     cuml/test/test_kmeans.py::test_all_kmeans_params[k-means||-32-0.5-100-10]
0.04s call     cuml/test/test_kmeans.py::test_all_kmeans_params[k-means||-32-0.5-500-10]
0.04s call     cuml/test/test_kmeans.py::test_all_kmeans_params[k-means||-32-1.5-100-10]
0.04s call     cuml/test/test_kmeans.py::test_all_kmeans_params[k-means||-32-1.5-500-10]
```

</details>